refactor: migrate authentication services to new structure and update user type definitions
