+++ src/services/ant-design-pro/login.ts
@@ -2,11 +2,11 @@
 /* eslint-disable */
 import { request } from '@umijs/max';
 
-/** 发送验证码 POST /api/login/captcha */
+/** Send verification code POST /api/login/captcha */
 export async function getFakeCaptcha(
   params: {
     // query
-    /** 手机号 */
+    /** Phone number */
     phone?: string;
   },
   options?: { [key: string]: any },
