+++ src/services/ant-design-pro/typings.d.ts
@@ -2,35 +2,6 @@
 /* eslint-disable */
 
 declare namespace API {
-  type CurrentUser = {
-    name?: string;
-    avatar?: string;
-    userid?: string;
-    teamid?: string;
-    email?: string;
-    signature?: string;
-    title?: string;
-    group?: string;
-    tags?: { key?: string; label?: string }[];
-    notifyCount?: number;
-    unreadCount?: number;
-    country?: string;
-    access?: string;
-    geographic?: {
-      province?: { label?: string; key?: string };
-      city?: { label?: string; key?: string };
-    };
-    address?: string;
-    phone?: string;
-    role?: string;
-  };
-
-  type LoginResult = {
-    status?: string;
-    type?: string;
-    currentAuthority?: string;
-  };
-
   type PageParams = {
     current?: number;
     pageSize?: number;
@@ -53,7 +24,7 @@ declare namespace API {
 
   type RuleList = {
     data?: RuleListItem[];
-    /** 列表的内容总数 */
+    /** Total number of list items */
     total?: number;
     success?: boolean;
   };
@@ -63,27 +34,18 @@ declare namespace API {
     status?: string;
   };
 
-  type LoginParams = {
-    email?: string;
-    username?: string;
-    password?: string;
-    confirmPassword?: string;
-    autoLogin?: boolean;
-    type?: string;
-  };
-
   type ErrorResponse = {
-    /** 业务约定的错误码 */
+    /** Business agreed error code */
     errorCode: string;
-    /** 业务上的错误信息 */
+    /** Business error message */
     errorMessage?: string;
-    /** 业务上的请求是否成功 */
+    /** Whether the business request was successful */
     success?: boolean;
   };
 
   type NoticeIconList = {
     data?: NoticeIconItem[];
-    /** 列表的内容总数 */
+    /** Total number of list items */
     total?: number;
     success?: boolean;
   };
