+++ src/services/auth/index.ts
@@ -1,0 +1,5 @@
+// Export all authentication related functions
+export { getCurrentUser, login, logout, reauthenticate, sendMagicLink, signUp } from './api';
+export { cognitoChangeEmail, cognitoChangePassword, cognitoSignUp } from './cognito';
+export { changePassword, forgotPasswordSendEmail, setPassword } from './password';
+export { changeEmail, setProfile } from './profile';
